Move GUID as a db field to its own file.
